move Model to its own module
